CLJ-1786: extend shouldn't change prot method impl
Previously, `extend` would bind _new_ protocol method
implementations. This meant every call to `extend` would
"unlink" previous implementations, meaning a deref of the method
var was required to see the new extensions.

This happened because the code generated by `emit-method-builder`
created a new method impls instead of sharing a single one for
the duration of the protocol's life.